+++ src/example/tag/tag.ts
@@ -1,6 +1,7 @@
 export interface Tag {
   readonly id: number;
   readonly name: string;
+  readonly ownerId: string;
 }
 
 export function tagToJSON(tag: Tag): { id: number; name: string } {